python_app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.0b1
pydantic-settings==2.3.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
SQLAlchemy==2.0.23
python-jose[cryptography]
slowapi==0.1.9
APScheduler==3.10.4
pytest==7.4.0
pytest-cov==4.1.0
alembic==1.11.1
structlog==23.1.0
flake8 